Fix dragged out UDL file is not applied to UDL in the new instance
The workaround is: in case of UDL, "-lLANG" argument part is ignored.
We let new instance detect the user lang type via file extension - it works in the most of case, except if user applies an UDL manually.
For example,  this workaround won't work under the following situation:
user applies Markdown to a file named "myMarkdown.abc".

Fix #3451