Display the accurate OS version in DebugInfo (for Windows build 21H1 and greater)
Query for DisplayVersion as well instead of using only ReleaseId.

Windows deprecated the ReleaseID registry entry (now it is no longer updated and on newer Windows 10/11 builds will remain the same) and moved to use DisplayVersion.
Attempt now to first query DisplayVersion and in case that is missing, fall back to ReleaseID for older Windows 10 builds.

Close #11714