Make session inaccessible files remembered (part 1/2)
Add the ability of "placrholders" (the empty document) for the inaccessible files of past session.

The inaccessible files from the current (default) session (i.e., session.xml) will not be lost in the next startup of Notepad++.
Furthermore, if file status detection is enabled in Notepad++ (default setting: ON), and the previously inaccessible files become accessible again (e.g., USB drive is inserted or network drive is remounted), the resurrected files can be reloaded on-the-fly without restarting Notepad++. The user will be prompted to reload if the placeholder is switched in.

Fix #12079, fix #12744, fix #13696, close #14168